ai-mock-interview-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL
alembic==1.12.1  # Database migrations

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# NLP
nltk==3.8.1
# spacy==3.7.2  # Optional, can be added later

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing async endpoints

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# For SQLite support (development)
sqlite3

# Additional useful packages
email-validator==2.1.0
python-json-logger==2.0.7